feat(DSAutoComplete): Add configurable ID field support
Allow customization of the ID field used in autocomplete component by adding
an optional 'idField' prop. This makes the component more flexible when working
with data sources where the id we're searching on is the pk for a different collection. Default behavior remains unchanged, using "id" as the default field name.