feat: Calculate overtime hours to pay considering banked hours
Updates the `new_time_report.sql` query to correctly calculate the "overtime hours to pay". For hourly employees, if overtime hours are designated to be banked (`overtimeHoursToBank > 0`), these are subtracted from the total calculated overtime hours (hours worked over 44) to determine the amount to be paid out. Otherwise, the total calculated overtime is used. Salaried employees have 0 overtime hours to pay according to this logic.